backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Talisman==1.1.0
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Security
bcrypt==4.1.1
cryptography==41.0.7

# AI/ML
openai==1.3.7
transformers==4.36.0
tensorflow==2.15.0
scikit-learn==1.3.2
torch==2.1.1

# AWS Services
boto3==1.34.0

# Task Scheduling
celery==5.3.4
APScheduler==3.10.4

# Validation and Serialization
marshmallow==3.20.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
textstat==0.7.3
